requirements.txt
antlr4-python3-runtime==4.9.3
bleach==6.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.0
hydra-core==1.3.2
idna==3.7
importlib-resources==6.4.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
omegaconf==2.3.0
packaging==24.1
pandas==2.0.3
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.12.1
text-unidecode==1.3
threadpoolctl==3.5.0
torch==2.2.2
torchsummary==1.5.1
torchvision==0.17.2
tqdm==4.66.4
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
zipp==3.19.2
